+++ App.tsx
@@ -9,9 +9,11 @@ import {
   useNavigationContainerRef,
 } from '@react-navigation/native';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import Dialog from 'components/common/Dialog/Dialog';
+import useDialog from 'hooks/app/useDialog';
 import AuthorizeNavigator from 'navigators/AuthorizeNavigator';
 import Navigator from 'navigators/Navigator';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
   SafeAreaView,
   StatusBar,
@@ -22,7 +24,6 @@ import {
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import SplashScreen from 'react-native-splash-screen';
 import { MyTheme, colors } from 'styles/theme';
-import Dialog from 'types/apps/dialog';
 import DialogContext from 'utils/DialogContext';
 
 const appStyles = StyleSheet.create({
@@ -53,33 +54,9 @@ if (__DEV__) {
 }
 
 const App = () => {
-  const [dialog, setDialog] = useState<Dialog>({
-    type: 'success',
-    text: '',
-    isShow: false,
-  });
+  const { dialogContextValue, dialog, closeDialog } = useDialog();
   const navigationRef = useNavigationContainerRef();
   useFlipper(navigationRef);
-
-  const openDialog = (text: string, type: string) => {
-    setDialog({
-      text,
-      type,
-      isShow: true,
-    });
-  };
-
-  const closeDialog = () => {
-    setDialog({
-      type: 'success',
-      text: '',
-      isShow: false,
-    });
-  };
-
-  const dialogContextValue = useMemo(() => {
-    return { openDialog, closeDialog };
-  }, [!!dialog.isShow]);
 
   const [isLogin, setIsLogin] = useState<boolean>(false);
 
@@ -94,12 +71,14 @@ const App = () => {
         <SafeAreaView style={appStyles.safeAreaContainer}>
           <GestureHandlerRootView style={appStyles.gestureContainer}>
             <NavigationContainer theme={MyTheme} ref={navigationRef}>
+              {/* <StorybookUIRoot /> */}
               <StatusBar backgroundColor={colors.background} />
               {isLogin ? (
                 <Navigator />
               ) : (
                 <AuthorizeNavigator setIsLogin={setIsLogin} />
               )}
+              <Dialog dialog={dialog} closeDialog={() => closeDialog} />
             </NavigationContainer>
           </GestureHandlerRootView>
         </SafeAreaView>
